Fix token scope tooltips on soft nav and place create-dialog tips beside titles.
Move TokenScopeBadges into a client module so TooltipTrigger asChild hydrates reliably, and open scope descriptions to the right of the label.
